requirements.txt
gym>=0.21.0
matplotlib==3.7.1
numpy==1.24.3
osmnx==1.1.2
pandas==1.5.3
parl==2.2.1
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
tensorboard>=2.11.0
torch>=1.13.1
tqdm==4.65.0
